Validaciones usuario

+++ usuarios.py
@@ -1,7 +1,27 @@
 import tkinter as tk
-from tkinter import ttk,END
+import re
+from tkinter import ttk,END, messagebox, simpledialog
 from backend.pefil import get_pefiles
-from backend.usuarios import post_usuario,get_usuario,update_usuario,delete_usuario,Usuario
+from backend.usuarios import post_usuario,get_usuario,update_usuario,delete_usuario,Usuario, get_max_id
+
+def validar_cadena(cadena):
+    if re.match('^[a-zA-Z\s]+$', cadena):
+        return cadena
+    else:
+        raise ValueError("Cadena tiene caracteres invalidos")
+
+def validar_email(email):
+    if re.match(r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$', email):
+        return email
+    else:
+        raise ValueError("Email no valido")
+
+def validar_lenght(cadena):
+    if len(cadena) >= 4:
+        return cadena
+    else:
+        raise ValueError("Cadena con menos de 4 caracteres")
+
 class Usuarios(tk.Frame):
     def __init__(self, parent):
         super().__init__(parent)
@@ -43,24 +63,50 @@ class Usuarios(tk.Frame):
         self.combo_usuarios_perfil.grid(row = 3, column = 3, padx = 10, pady = 10)
         self.combo_usuarios_perfil.config(values=self.perfiles)
         # Buttons
-        button_usuarios_buscar = tk.Button(self, text = "Buscar")
-        button_usuarios_nuevo = tk.Button(self, text = "Nuevo")
-        button_usuarios_guardar = tk.Button(self, text = "Guardar")
-        button_usuarios_cancelar = tk.Button(self, text = "Cancelar")
+        self.button_usuarios_buscar = tk.Button(self, text = "Buscar", command=lambda:self.buscar_usuario())
+        self.button_usuarios_nuevo = tk.Button(self, text = "Nuevo", command = lambda:self.nuevo_usuario())
+        self.button_usuarios_guardar = tk.Button(self, text = "Guardar", command=lambda:self.guardar_usuario())
+        self.button_usuarios_cancelar = tk.Button(self, text = "Cancelar", command=lambda:self.cancelar_usuario())
         self.button_usuarios_editar = tk.Button(self, text = "Editar")
         self.button_usuarios_baja = tk.Button(self, text = "Baja")
-        button_usuarios_buscar.grid(row = 0, column = 2, padx = 10, pady = 10)
-        button_usuarios_nuevo.grid(row = 6, column = 0, padx = 10, pady = 10)
-        button_usuarios_guardar.grid(row = 6, column = 1, padx = 10, pady = 10)
-        button_usuarios_cancelar.grid(row = 6, column = 2, padx = 10, pady = 10)
+        self.button_usuarios_buscar.grid(row = 0, column = 2, padx = 10, pady = 10)
+        self.button_usuarios_nuevo.grid(row = 6, column = 0, padx = 10, pady = 10)
+        self.button_usuarios_guardar.grid(row = 6, column = 1, padx = 10, pady = 10)
+        self.button_usuarios_cancelar.grid(row = 6, column = 2, padx = 10, pady = 10)
         self.button_usuarios_editar.grid(row = 6, column = 3, padx = 10, pady = 10)
         self.button_usuarios_baja.grid(row = 6, column = 4, padx = 10, pady = 10)
 
-        button_usuarios_guardar.config(command=lambda:self.guardar_usuario())
+        self.entry_usuarios_id.config(state = "disabled")
 
-        button_usuarios_buscar.config(command=lambda:self.buscar_usuario())
-        
-        button_usuarios_cancelar.config(command=lambda:self.limpiar_campos())
+        self.button_usuarios_nuevo.config(state = "active")
+        self.button_usuarios_buscar.config(state = "active")
+        self.button_usuarios_guardar.config(state = "disabled")
+        self.button_usuarios_cancelar.config(state = "disabled")
+        self.button_usuarios_editar.config(state = "disabled")
+        self.button_usuarios_baja.config(state = "disabled")
+
+    def nuevo_usuario(self):
+        self.button_usuarios_nuevo.config(state = "disabled")
+        self.button_usuarios_buscar.config(state = "disabled")
+        self.button_usuarios_guardar.config(state = "active")
+        self.button_usuarios_cancelar.config(state = "active")
+        self.button_usuarios_editar.config(state = "disabled")
+        self.button_usuarios_baja.config(state = "disabled")
+        self.entry_usuarios_id.config(state = "normal")
+        self.limpiar_campos()
+        self.entry_usuarios_id.insert(0, int(get_max_id()[0]) + 1)
+        self.entry_usuarios_id.config(state = "disabled")
+
+    def cancelar_usuario(self):
+        self.entry_usuarios_id.config(state = "normal")
+        self.limpiar_campos()
+        self.entry_usuarios_id.config(state = "disabled")
+        self.button_usuarios_nuevo.config(state = "active")
+        self.button_usuarios_buscar.config(state = "active")
+        self.button_usuarios_guardar.config(state = "disabled")
+        self.button_usuarios_cancelar.config(state = "disabled")
+        self.button_usuarios_editar.config(state = "disabled")
+        self.button_usuarios_baja.config(state = "disabled")
 
     def representar_perfil(self,seleccion):
         for x in self.perfiles:
@@ -69,15 +115,21 @@ class Usuarios(tk.Frame):
         return None
     
     def guardar_usuario(self):
-        post_usuario(
-            self.representar_perfil(self.combo_usuarios_perfil.get()),
-            self.entry_usuarios_nombre.get(),
-            self.entry_usuarios_paterno.get(),
-            self.entry_usuarios_materno.get(),
-            self.entry_usuarios_email.get(),
-            self.entry_usuarios_password.get(),
-            self.entry_usuarios_username.get(),
-        )
+        try:
+            post_usuario(
+                self.representar_perfil(self.combo_usuarios_perfil.get()),
+                validar_cadena(self.entry_usuarios_nombre.get()),
+                validar_cadena(self.entry_usuarios_paterno.get()),
+                validar_cadena(self.entry_usuarios_materno.get()),
+                validar_email(self.entry_usuarios_email.get()),
+                validar_lenght(self.entry_usuarios_password.get()),
+                validar_lenght(self.entry_usuarios_username.get()),
+            )
+            messagebox.showinfo(title = "Aviso", message = "Usuario guardado con exito")
+            self.cancelar_usuario()
+        except ValueError:
+            messagebox.showerror(title = "Error", message = "Datos invalidos o incompletos")
+        
     def colocar_datos_en_entrys(self,usuario):
         self.entry_usuarios_id.insert(0, usuario.id)
         self.entry_usuarios_nombre.insert(0, usuario.nombre)
@@ -99,22 +151,43 @@ class Usuarios(tk.Frame):
         self.combo_usuarios_perfil.delete(0,END)
 
     def editar_usuario(self,id):
-        usuario =Usuario((id,self.representar_perfil(self.combo_usuarios_perfil.get()),
-            self.entry_usuarios_nombre.get(),
-            self.entry_usuarios_paterno.get(),
-            self.entry_usuarios_materno.get(),
-            self.entry_usuarios_email.get(),
-            self.entry_usuarios_password.get(),
-            self.entry_usuarios_username.get(),
-            'ACTIVO')) 
-        update_usuario(id,usuario)
+        try:
+            usuario =Usuario((id,self.representar_perfil(self.combo_usuarios_perfil.get()),
+                validar_cadena(self.entry_usuarios_nombre.get()),
+                validar_cadena(self.entry_usuarios_paterno.get()),
+                validar_cadena(self.entry_usuarios_materno.get()),
+                validar_email(self.entry_usuarios_email.get()),
+                validar_lenght(self.entry_usuarios_password.get()),
+                validar_lenght(self.entry_usuarios_username.get()),
+                'ACTIVO')) 
+            update_usuario(id,usuario)
+            messagebox.showinfo(title = "Aviso", message = "Usuario editado con exito")
+            self.cancelar_usuario()
+        except:
+            messagebox.showerror(title = "Error", message = "Datos invalidos o incompletos")
+        
 
     def borrar_usuario(self,id):
-        delete_usuario(id)
+        confirmar = messagebox.askyesno(title = "Confirmacion", message = f"¿Estás seguro de eliminar al usuario con el id: {id} ?")
+        if confirmar:
+            delete_usuario(id)
+            messagebox.showinfo(title = "Aviso", message = "Usuario eliminado con exito")
+            self.cancelar_usuario()
+        else:
+            messagebox.showerror(title = "Error", message = "Operacion cancelada")
 
     def buscar_usuario(self):
-        self.limpiar_campos()
-        usuario = Usuario(get_usuario(self.entry_usuarios_buscar.get()))
-        self.colocar_datos_en_entrys(usuario)
-        self.button_usuarios_editar.config(command=lambda:self.editar_usuario(usuario.id))
-        self.button_usuarios_baja.config(command=lambda:self.borrar_usuario(usuario.id))
+        try:
+            usuario = Usuario(get_usuario(int(self.entry_usuarios_buscar.get())))
+            self.limpiar_campos()
+            self.colocar_datos_en_entrys(usuario)
+            self.button_usuarios_editar.config(command=lambda:self.editar_usuario(usuario.id))
+            self.button_usuarios_baja.config(command=lambda:self.borrar_usuario(usuario.id))
+            self.button_usuarios_nuevo.config(state = "disabled")
+            self.button_usuarios_buscar.config(state = "active")
+            self.button_usuarios_guardar.config(state = "disabled")
+            self.button_usuarios_cancelar.config(state = "active")
+            self.button_usuarios_editar.config(state = "active")
+            self.button_usuarios_baja.config(state = "active")
+        except:
+            messagebox.showerror(title = "Error", message = "Dato de busqueda invalido")
